Struct google_cloudidentity1::GroupMethods

pub struct GroupMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on group resources. It is not used directly, but through the CloudIdentity hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_cloudidentity1 as cloudidentity1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use cloudidentity1::CloudIdentity;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = CloudIdentity::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `create(...)`, `delete(...)`, `get(...)`, `list(...)`, `lookup(...)`, `memberships_create(...)`, `memberships_delete(...)`, `memberships_get(...)`, `memberships_list(...)`, `memberships_lookup(...)`, `patch(...)` and `search(...)`
// to build up your call.
let rb = hub.groups();

Methods

impl<'a, C, A> GroupMethods<'a, C, A>

pub fn memberships_create(
&self,
request: Membership,
parent: &str
) -> GroupMembershipCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a Membership.

Arguments

request - No description provided.

parent - Resource name of the Group to create Membership within. Format: groups/{group_id}, where group_id is the unique ID assigned to the Group.

pub fn memberships_list(
&self,
parent: &str
) -> GroupMembershipListCall<'a, C, A>

Create a builder to help you perform the following task:

List Memberships within a Group.

Arguments

parent - Resource name of the Group to list Memberships within. Format: groups/{group_id}, where group_id is the unique ID assigned to the Group.

pub fn list(&self) -> GroupListCall<'a, C, A>

Create a builder to help you perform the following task:

List groups within a customer or a domain.

pub fn get(&self, name: &str) -> GroupGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a Group.

Arguments

name - Resource name of the Group in the format: groups/{group_id}, where group_id is the unique ID assigned to the Group.

pub fn memberships_delete(
&self,
name: &str
) -> GroupMembershipDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a Membership.

Arguments

name - Resource name of the Membership to be deleted. Format: groups/{group_id}/memberships/{member_id}, where group_id is the unique ID assigned to the Group to which Membership belongs to, and member_id is the unique ID assigned to the member.

pub fn search(&self) -> GroupSearchCall<'a, C, A>

Create a builder to help you perform the following task:

Searches for Groups.

pub fn memberships_get(&self, name: &str) -> GroupMembershipGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a Membership.

Arguments

name - Resource name of the Membership to be retrieved. Format: groups/{group_id}/memberships/{member_id}, where group_id is the unique id assigned to the Group to which Membership belongs to, and member_id is the unique ID assigned to the member.

pub fn create(&self, request: Group) -> GroupCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a Group.

Arguments

request - No description provided.

pub fn lookup(&self) -> GroupLookupCall<'a, C, A>

Create a builder to help you perform the following task:

Looks up resource name of a Group by its EntityKey.

pub fn delete(&self, name: &str) -> GroupDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a Group.

Arguments

name - Resource name of the Group in the format: groups/{group_id}, where group_id is the unique ID assigned to the Group.

pub fn patch(&self, request: Group, name: &str) -> GroupPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a Group.

Arguments

request - No description provided.

name - Resource name of the Group in the format: groups/{group_id}, where group_id is the unique ID assigned to the Group. Must be left blank while creating a Group.

pub fn memberships_lookup(
&self,
parent: &str
) -> GroupMembershipLookupCall<'a, C, A>

Create a builder to help you perform the following task:

Looks up resource name of a Membership within a Group by member's EntityKey.

Arguments

parent - Resource name of the Group to lookup Membership within. Format: groups/{group_id}, where group_id is the unique ID assigned to the Group.