Struct google_cloudidentity1::api::GroupMethods

pub struct GroupMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on group resources. It is not used directly, but through the CloudIdentity hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_cloudidentity1 as cloudidentity1;
 
use std::default::Default;
use oauth2;
use cloudidentity1::CloudIdentity;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = CloudIdentity::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `create(...)`, `delete(...)`, `get(...)`, `list(...)`, `lookup(...)`, `memberships_check_transitive_membership(...)`, `memberships_create(...)`, `memberships_delete(...)`, `memberships_get(...)`, `memberships_get_membership_graph(...)`, `memberships_list(...)`, `memberships_lookup(...)`, `memberships_modify_membership_roles(...)`, `memberships_search_transitive_groups(...)`, `memberships_search_transitive_memberships(...)`, `patch(...)` and `search(...)`
// to build up your call.
let rb = hub.groups();

Implementations

impl<'a, C> GroupMethods<'a, C>

pub fn memberships_check_transitive_membership(
&self,
parent: &str
) -> GroupMembershipCheckTransitiveMembershipCall<'a, C>

Create a builder to help you perform the following task:

Check a potential member for membership in a group. Note: This feature is only available to Google Workspace Enterprise Standard, Enterprise Plus, and Enterprise for Education; and Cloud Identity Premium accounts. If the account of the member is not one of these, a 403 (PERMISSION_DENIED) HTTP status code will be returned. A member has membership to a group as long as there is a single viewable transitive membership between the group and the member. The actor must have view permissions to at least one transitive membership between the member and group.

Arguments

parent - Resource name of the group to check the transitive membership in. Format: groups/{group_id}, where group_id is the unique id assigned to the Group to which the Membership belongs to.

pub fn memberships_create(
&self,
request: Membership,
parent: &str
) -> GroupMembershipCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a Membership.

Arguments

request - No description provided.

parent - Required. The parent Group resource under which to create the Membership. Must be of the form groups/{group_id}.

pub fn memberships_delete(&self, name: &str) -> GroupMembershipDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a Membership.

Arguments

name - Required. The resource name of the Membership to delete. Must be of the form groups/{group_id}/memberships/{membership_id}

pub fn memberships_get(&self, name: &str) -> GroupMembershipGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves a Membership.

Arguments

name - Required. The resource name of the Membership to retrieve. Must be of the form groups/{group_id}/memberships/{membership_id}.

pub fn memberships_get_membership_graph(
&self,
parent: &str
) -> GroupMembershipGetMembershipGraphCall<'a, C>

Create a builder to help you perform the following task:

Get a membership graph of just a member or both a member and a group. Note: This feature is only available to Google Workspace Enterprise Standard, Enterprise Plus, and Enterprise for Education; and Cloud Identity Premium accounts. If the account of the member is not one of these, a 403 (PERMISSION_DENIED) HTTP status code will be returned. Given a member, the response will contain all membership paths from the member. Given both a group and a member, the response will contain all membership paths between the group and the member.

Arguments

parent - Required. Resource name of the group to search transitive memberships in. Format: groups/{group_id}, where group_id is the unique ID assigned to the Group to which the Membership belongs to. group_id can be a wildcard collection id "-". When a group_id is specified, the membership graph will be constrained to paths between the member (defined in the query) and the parent. If a wildcard collection is provided, all membership paths connected to the member will be returned.

pub fn memberships_list(&self, parent: &str) -> GroupMembershipListCall<'a, C>

Create a builder to help you perform the following task:

Lists the Memberships within a Group.

Arguments

parent - Required. The parent Group resource under which to lookup the Membership name. Must be of the form groups/{group_id}.

pub fn memberships_lookup(
&self,
parent: &str
) -> GroupMembershipLookupCall<'a, C>

Create a builder to help you perform the following task:

Looks up the resource name of a Membership by its EntityKey.

Arguments

parent - Required. The parent Group resource under which to lookup the Membership name. Must be of the form groups/{group_id}.

pub fn memberships_modify_membership_roles(
&self,
request: ModifyMembershipRolesRequest,
name: &str
) -> GroupMembershipModifyMembershipRoleCall<'a, C>

Create a builder to help you perform the following task:

Modifies the MembershipRoles of a Membership.

Arguments

request - No description provided.

name - Required. The resource name of the Membership whose roles are to be modified. Must be of the form groups/{group_id}/memberships/{membership_id}.

pub fn memberships_search_transitive_groups(
&self,
parent: &str
) -> GroupMembershipSearchTransitiveGroupCall<'a, C>

Create a builder to help you perform the following task:

Search transitive groups of a member. Note: This feature is only available to Google Workspace Enterprise Standard, Enterprise Plus, and Enterprise for Education; and Cloud Identity Premium accounts. If the account of the member is not one of these, a 403 (PERMISSION_DENIED) HTTP status code will be returned. A transitive group is any group that has a direct or indirect membership to the member. Actor must have view permissions all transitive groups.

Arguments

parent - Resource name of the group to search transitive memberships in. Format: groups/{group_id}, where group_id is always '-' as this API will search across all groups for a given member.

pub fn memberships_search_transitive_memberships(
&self,
parent: &str
) -> GroupMembershipSearchTransitiveMembershipCall<'a, C>

Create a builder to help you perform the following task:

Search transitive memberships of a group. Note: This feature is only available to Google Workspace Enterprise Standard, Enterprise Plus, and Enterprise for Education; and Cloud Identity Premium accounts. If the account of the group is not one of these, a 403 (PERMISSION_DENIED) HTTP status code will be returned. A transitive membership is any direct or indirect membership of a group. Actor must have view permissions to all transitive memberships.

Arguments

parent - Resource name of the group to search transitive memberships in. Format: groups/{group_id}, where group_id is the unique ID assigned to the Group.

pub fn create(&self, request: Group) -> GroupCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a Group.

Arguments

request - No description provided.

pub fn delete(&self, name: &str) -> GroupDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a Group.

Arguments

name - Required. The resource name of the Group to retrieve. Must be of the form groups/{group_id}.

pub fn get(&self, name: &str) -> GroupGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves a Group.

Arguments

name - Required. The resource name of the Group to retrieve. Must be of the form groups/{group_id}.

pub fn list(&self) -> GroupListCall<'a, C>

Create a builder to help you perform the following task:

Lists the Groups under a customer or namespace.

pub fn lookup(&self) -> GroupLookupCall<'a, C>

Create a builder to help you perform the following task:

Looks up the resource name of a Group by its EntityKey.

pub fn patch(&self, request: Group, name: &str) -> GroupPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a Group.

Arguments

request - No description provided.

name - Output only. The resource name of the Group. Shall be of the form groups/{group_id}.

pub fn search(&self) -> GroupSearchCall<'a, C>

Create a builder to help you perform the following task:

Searches for Groups matching a specified query.